pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-own-jarvis"
version = "0.1.4"
description = "Fully local voice assistant inspired by J.A.R.V.I.S."
readme = { file = "README_PYPI.md", content-type = "text/markdown" }
requires-python = ">=3.11,<3.13"
authors = [{ name = "lusparkl" }]
keywords = ["jarvis", "voice-assistant", "local-ai", "ollama", "speech"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "faster-whisper",
    "coqui-tts[codec]==0.27.5",
    "transformers==4.57.6",
    "torch==2.10.0",
    "torchaudio==2.10.0",
    "torchvision==0.25.0",
    "sounddevice",
    "openwakeword",
    "python-dotenv",
    "ollama",
    "chromadb",
    "pyperclip",
    "numpy",
    "scipy",
    "requests",
]

[project.urls]
Homepage = "https://github.com/lusparkl/local_ai_assistant_jarvis"
Repository = "https://github.com/lusparkl/local_ai_assistant_jarvis"
Issues = "https://github.com/lusparkl/local_ai_assistant_jarvis/issues"

[project.scripts]
jarvis = "jarvis_cli:main"

[tool.setuptools]
include-package-data = true
py-modules = ["main", "config", "jarvis_cli", "preflight"]

[tool.setuptools.packages.find]
include = ["audio*", "models*", "services*", "tools*", "setup*"]
namespaces = true

[tool.setuptools.package-data]
audio = ["pointing_sounds/*.wav", "reference_sounds/*.wav", "thinking_sounds/*.wav"]
